refactor(all): complete most core service logic refactor
shared user services are pulled to core service and many auth are checked in core logic. Require
some more work to translate services for other entities if needed but current implementation seems
reasonable. The framework is also more or less easy to work with right now. Might need to break down
the core files though if we were to move all CRUD over.

Closes #7.